+++ src/common/execution/order-helper.js
@@ -8,9 +8,11 @@ import {
   NULL_ADDRESS,
   NULL_BYTES32,
   IEXEC_REQUEST_PARAMS,
+  STORAGE_PROVIDERS,
   TEE_FRAMEWORKS,
 } from '../utils/constant.js';
 import {
+  getStorageTokenKeyName,
   reservedSecretKeyName,
 } from '../utils/secrets-utils.js';
 import {
@@ -90,6 +92,28 @@ export const checkRequestRequirements = async (
     if (!isEncryptionKeySet) {
       throw Error(
         'Beneficiary result encryption key is not set in the SMS. Result encryption will fail.',
+      );
+    }
+  }
+
+  // check dropbox storage token
+  if (
+    paramsObj[IEXEC_REQUEST_PARAMS.IEXEC_RESULT_STORAGE_PROVIDER] ===
+      STORAGE_PROVIDERS.DROPBOX
+  ) {
+    const isStorageTokenSet = await checkWeb2SecretExists(
+      contracts,
+      smsURL,
+      requester,
+      getStorageTokenKeyName(
+        paramsObj[IEXEC_REQUEST_PARAMS.IEXEC_RESULT_STORAGE_PROVIDER],
+      ),
+    );
+    if (!isStorageTokenSet) {
+      throw Error(
+        `Requester storage token is not set for selected provider "${
+          paramsObj[IEXEC_REQUEST_PARAMS.IEXEC_RESULT_STORAGE_PROVIDER]
+        }". Result archive upload will fail.`,
       );
     }
   }